refactor(git): split git_operations.py into git/ package
Split the 460-LOC git_operations module into focused submodules:
- git/validation.py: URL validation, sanitization, transient error detection
- git/operations.py: core git CLI runner, branch/commit/push/diff
- git/clone.py: git clone with retry logic
- git/patches.py: patch validation and application
- git/archive.py: tarball creation and extraction
- git/__init__.py: re-exports all public symbols

git_operations.py remains as a backward-compatible re-export shim for
tests that import private helpers. All source imports updated to use
the new git package directly.
